+++ docs/versionhistory.rst
@@ -18,6 +18,7 @@ APScheduler, see the :doc:`migration section <migration>`.
 - Fixed large parts of ``MongoDBDataStore`` still calling blocking functions in the
   event loop thread
 - Fixed JSON serialization of triggers that had been used at least once
+- Fixed dialect name checks in the SQLAlchemy job store
 
 **4.0.0a4**